Complete, ideally correct quantization implementation. All tests passing.
This one got pretty tricky but I finally got the inspiration I needed to
solve it. Previous implementation was only able to quantize 'balanced'
domains e.g. [-20, 20]. It failed for domains like [-25, 10]. This was
because the 'distance' algorithm couldn't compensate for the
'unbalanced' nature of the domains.

An if/else in quantize and the addition of balanceSteps creates a
'shifted' version of the steps array which is used for distance
calculations. This works because the shifted array's indices still
correlate to the original steps.